wb_fbs/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile
aioboto3==9.6.0
    # via -r requirements.in
aiobotocore[boto3]==2.3.0
    # via aioboto3
aiohttp==3.8.1
    # via aiobotocore
aioitertools==0.10.0
    # via aiobotocore
aiosignal==1.2.0
    # via aiohttp
anyio==3.6.1
    # via
    #   httpcore
    #   starlette
async-timeout==4.0.2
    # via aiohttp
attrs==22.1.0
    # via aiohttp
boto3==1.21.21
    # via aiobotocore
botocore==1.24.21
    # via
    #   aiobotocore
    #   boto3
    #   s3transfer
certifi==2022.6.15
    # via
    #   httpcore
    #   httpx
    #   sentry-sdk
charset-normalizer==2.1.0
    # via aiohttp
click==8.1.3
    # via uvicorn
deprecation==2.1.0
    # via pikepdf
dnspython==2.2.1
    # via pymongo
fastapi==0.79.0
    # via
    #   -r requirements.in
    #   fastapi-utils
fastapi-utils==0.2.1
    # via -r requirements.in
frozenlist==1.3.1
    # via
    #   aiohttp
    #   aiosignal
greenlet==1.1.2
    # via sqlalchemy
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via httpx
httpx==0.23.0
    # via -r requirements.in
idna==3.3
    # via
    #   anyio
    #   rfc3986
    #   yarl
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lxml==4.9.1
    # via pikepdf
motor==3.0.0
    # via -r requirements.in
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
packaging==21.3
    # via
    #   deprecation
    #   pikepdf
pikepdf==5.4.2
    # via -r requirements.in
pillow==9.2.0
    # via
    #   pikepdf
    #   reportlab
pydantic[dotenv]==1.9.1
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-utils
pymongo[srv]==4.2.0
    # via
    #   -r requirements.in
    #   motor
pymupdf==1.20.1
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
pypdf3==1.0.6
    # via -r requirements.in
python-dateutil==2.8.2
    # via botocore
python-dotenv==0.20.0
    # via pydantic
pyyaml==6.0
    # via -r requirements.in
reportlab==3.6.11
    # via -r requirements.in
rfc3986[idna2008]==1.5.0
    # via httpx
s3transfer==0.5.2
    # via boto3
sentry-sdk==1.9.2
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.2.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==1.4.39
    # via fastapi-utils
starlette==0.19.1
    # via fastapi
tqdm==4.64.0
    # via pypdf3
typing-extensions==4.3.0
    # via
    #   aioitertools
    #   pydantic
    #   starlette
urllib3==1.26.11
    # via
    #   botocore
    #   sentry-sdk
uvicorn==0.18.2
    # via -r requirements.in
wrapt==1.14.1
    # via aiobotocore
yarl==1.8.1
    # via aiohttp
